Restructure directory and adjust test files to have tokenizers in src
